Move src/celery under src/shared
