+++ src/main/queue.ts
@@ -2,7 +2,7 @@ import Piscina from '@posthog/piscina'
 import { PluginEvent } from '@posthog/plugin-scaffold'
 import * as Sentry from '@sentry/node'
 
-import Client from '../celery/client'
+import Client from '../shared/celery/client'
 import { status } from '../status'
 import { IngestEventResponse, PluginsServer, Queue } from '../types'
 import { UUIDT } from '../utils'
