+++ src/worker/ingestion/process-event.ts
@@ -7,11 +7,11 @@ import { DateTime, Duration } from 'luxon'
 import * as fetch from 'node-fetch'
 import { nodePostHog } from 'posthog-js-lite/dist/src/targets/node'
 
-import Client from '../../celery/client'
 import { DB } from '../../db'
 import { Event as EventProto, IEvent } from '../../idl/protos'
 import { KAFKA_EVENTS, KAFKA_SESSION_RECORDING_EVENTS } from '../../ingestion/topics'
 import { elementsToString, personInitialAndUTMProperties, sanitizeEventName, timeoutGuard } from '../../ingestion/utils'
+import Client from '../../shared/celery/client'
 import { status } from '../../status'
 import {
     CohortPeople,
