+++ tests/postgres/worker.test.ts
@@ -2,9 +2,9 @@ import { PluginEvent } from '@posthog/plugin-scaffold/src/types'
 import IORedis from 'ioredis'
 import { mocked } from 'ts-jest/utils'
 
-import Client from '../../src/celery/client'
 import { loadPluginSchedule } from '../../src/main/services/schedule'
 import { ServerInstance, startPluginsServer } from '../../src/server'
+import Client from '../../src/shared/celery/client'
 import { LogLevel } from '../../src/types'
 import { delay, UUIDT } from '../../src/utils'
 import { ingestEvent } from '../../src/worker/ingestion/ingest-event'
